chore(storybook): refactor preview from js to ts and use ESM format
