+++ libs/storybook/.storybook/tsconfig.json
@@ -1,7 +1,8 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
-    "emitDecoratorMetadata": true
+    "emitDecoratorMetadata": true,
+    "resolveJsonModule": true
   },
   "exclude": ["../**/*.spec.ts", "jest.config.ts"],
   "include": [
@@ -14,6 +15,7 @@
     "*.js"
   ],
   "files": [
+    "preview.ts",
     "../../../globals.d.ts",
     "../../ui/src/index.ts",
     "../src/lib/storybook.module.ts"
